[Feature] Admin Controls

+++ backend/src/main/java/com/shopkart/user/dto/request/UserDetailsRequest.java
@@ -19,7 +19,7 @@ public class UserDetailsRequest {
     private String phoneNumber;
 
     @JsonProperty("gender")
-    private Short gender;
+    private Integer gender;
 
     @JsonProperty("date_of_birth")
     private LocalDate dateOfBirth;
