+++ backend/src/main/java/com/shopkart/user/dto/response/UserDetailsResponse.java
@@ -21,7 +21,7 @@ public class UserDetailsResponse {
     private String phoneNumber;
 
     @JsonProperty("gender")
-    private Short gender;
+    private Integer gender;
 
     @JsonProperty("date_of_birth")
     private LocalDate dateOfBirth;
